# skipPS option for CorrectedECALPFClusterProducer