--- a/RecoParticleFlow/PFClusterProducer/plugins/CorrectedECALPFClusterProducer.cc
+++ b/RecoParticleFlow/PFClusterProducer/plugins/CorrectedECALPFClusterProducer.cc
@@ -44,12 +44,13 @@ namespace {
 class CorrectedECALPFClusterProducer : public edm::stream::EDProducer<> {
 public:
   CorrectedECALPFClusterProducer(const edm::ParameterSet& conf)
-      : _minimumPSEnergy(conf.getParameter<double>("minimumPSEnergy")) {
+      : _minimumPSEnergy(conf.getParameter<double>("minimumPSEnergy")), _skipPS(conf.getParameter<bool>("skipPS")) {
     const edm::InputTag& inputECAL = conf.getParameter<edm::InputTag>("inputECAL");
     _inputECAL = consumes<reco::PFClusterCollection>(inputECAL);
 
     const edm::InputTag& inputPS = conf.getParameter<edm::InputTag>("inputPS");
-    _inputPS = consumes<reco::PFClusterCollection>(inputPS);
+    if (!_skipPS)
+      _inputPS = consumes<reco::PFClusterCollection>(inputPS);
 
     const edm::ParameterSet& corConf = conf.getParameterSet("energyCorrector");
     _corrector = std::make_unique<PFClusterEMEnergyCorrector>(corConf, consumesCollector());
@@ -63,6 +64,7 @@ public:
 
 private:
   const double _minimumPSEnergy;
+  const bool _skipPS;
   std::unique_ptr<PFClusterEMEnergyCorrector> _corrector;
   edm::EDGetTokenT<reco::PFClusterCollection> _inputECAL;
   edm::EDGetTokenT<reco::PFClusterCollection> _inputPS;
@@ -77,41 +79,45 @@ void CorrectedECALPFClusterProducer::produce(edm::Event& e, const edm::EventSetu
   edm::Handle<reco::PFClusterCollection> handleECAL;
   e.getByToken(_inputECAL, handleECAL);
   edm::Handle<reco::PFClusterCollection> handlePS;
-  e.getByToken(_inputPS, handlePS);
+  if (!_skipPS)
+    e.getByToken(_inputPS, handlePS);
 
   auto const& ecals = *handleECAL;
-  auto const& pss = *handlePS;
 
   clusters_out->reserve(ecals.size());
   association_out->reserve(ecals.size());
   clusters_out->insert(clusters_out->end(), ecals.begin(), ecals.end());
+
   //build the EE->PS association
-  for (unsigned i = 0; i < pss.size(); ++i) {
-    switch (pss[i].layer()) {  // just in case this isn't the ES...
-      case PFLayer::PS1:
-      case PFLayer::PS2:
-        break;
-      default:
-        continue;
-    }
-    if (pss[i].energy() < _minimumPSEnergy)
-      continue;
-    int eematch = -1;
-    auto min_dist = std::numeric_limits<double>::max();
-    for (size_t ic = 0; ic < ecals.size(); ++ic) {
-      if (ecals[ic].layer() != PFLayer::ECAL_ENDCAP)
-        continue;
-      auto dist = testPreshowerDistance(ecals[ic], pss[i]);
-      if (dist == -1.0)
-        dist = std::numeric_limits<double>::max();
-      if (dist < min_dist) {
-        eematch = ic;
-        min_dist = dist;
+  if (!_skipPS) {
+    auto const& pss = *handlePS;
+    for (unsigned i = 0; i < pss.size(); ++i) {
+      switch (pss[i].layer()) {  // just in case this isn't the ES...
+        case PFLayer::PS1:
+        case PFLayer::PS2:
+          break;
+        default:
+          continue;
       }
-    }  // loop on EE clusters
-    if (eematch >= 0) {
-      edm::Ptr<reco::PFCluster> psclus(handlePS, i);
-      association_out->push_back(std::make_pair(eematch, psclus));
+      if (pss[i].energy() < _minimumPSEnergy)
+        continue;
+      int eematch = -1;
+      auto min_dist = std::numeric_limits<double>::max();
+      for (size_t ic = 0; ic < ecals.size(); ++ic) {
+        if (ecals[ic].layer() != PFLayer::ECAL_ENDCAP)
+          continue;
+        auto dist = testPreshowerDistance(ecals[ic], pss[i]);
+        if (dist == -1.0)
+          dist = std::numeric_limits<double>::max();
+        if (dist < min_dist) {
+          eematch = ic;
+          min_dist = dist;
+        }
+      }  // loop on EE clusters
+      if (eematch >= 0) {
+        edm::Ptr<reco::PFCluster> psclus(handlePS, i);
+        association_out->push_back(std::make_pair(eematch, psclus));
+      }
     }
   }
   std::sort(association_out->begin(), association_out->end(), sortByKey);
@@ -127,6 +133,7 @@ void CorrectedECALPFClusterProducer::produce(edm::Event& e, const edm::EventSetu
 void CorrectedECALPFClusterProducer::fillDescriptions(edm::ConfigurationDescriptions& descriptions) {
   edm::ParameterSetDescription desc;
   desc.add<double>("minimumPSEnergy", 0.0);
+  desc.add<bool>("skipPS", false);
   desc.add<edm::InputTag>("inputPS", edm::InputTag("particleFlowClusterPS"));
   {
     edm::ParameterSetDescription psd0;