@hatakeyamak 2021-06-16
Given the discussion in today's TSG about this option, it will be nice to get the test going.
https://indico.cern.ch/event/1040953/#35-pog-status-reports-particle