@hatakeyamak 2021-06-17
ok. no difference with reco is expected because we didn't change the default behavior. This is just introducing a switch we can test mainly for HLT purpose.